+++ join/h2.cpp
@@ -1,24 +1,24 @@
 #include "h2.hpp"
 
-int32_t h2(int32_t value, uint32_t prime){
-    // uint32_t prime = nextPrime(start);
+uint64_t h2(uint64_t value, uint64_t prime){
+    // uint64_t prime = nextPrime(start);
     return value % prime;
 }
 
-void testNextPrime(uint32_t upTo){
+void testNextPrime(uint64_t upTo){
     std::cout << "Testing nextPrime up to " << upTo << std::endl;
-    for(uint32_t i=0; i<upTo; i++){
+    for(uint64_t i=0; i<upTo; i++){
         std::cout << i << ". " << nextPrime(i) << std::endl;
     }
 }
 
-uint32_t nextPrime(uint32_t start){
+uint64_t nextPrime(uint64_t start){
     // Make sure we start checking from an odd number larger than 'start'
     if(start%2 == 0) start+=1;
     else start+=2;
 
     // Variable 'i' will be our next potential prime
-    uint32_t i = start;
+    uint64_t i = start;
 
     while(1){
         if(isPrime(i)) break;
@@ -31,8 +31,8 @@ uint32_t nextPrime(uint32_t start){
     return i;
 }
 
-bool isPrime(uint32_t i){
-    uint32_t d = 3; // First possible divisor
+bool isPrime(uint64_t i){
+    uint64_t d = 3; // First possible divisor
     bool flag = 1;  // Flag to stop the search if we find a divisor
 
     // While the divisor is less than the square root of out the number
@@ -52,40 +52,40 @@ bool isPrime(uint32_t i){
 }
 
 void bucketify2(Tuple * rel,
-                uint32_t bucketSize,
-                uint32_t startingPos,
-                uint32_t ** bucketArray,
-                uint32_t ** chainArray) {
+                uint64_t bucketSize,
+                uint64_t startingPos,
+                uint64_t ** bucketArray,
+                uint64_t ** chainArray) {
 
-    uint32_t prime = nextPrime(bucketSize);
+    uint64_t prime = nextPrime(bucketSize);
 
-    *bucketArray = new uint32_t[prime];
-    *chainArray = new uint32_t[bucketSize];
+    *bucketArray = new uint64_t[prime];
+    *chainArray = new uint64_t[bucketSize];
 
     // Initialise our arrays
-    for (uint32_t j = 0; j < bucketSize; j++) {
+    for (uint64_t j = 0; j < bucketSize; j++) {
         (*chainArray)[j] = -1;
     }
-    for (uint32_t j = 0; j < prime; j++) {
+    for (uint64_t j = 0; j < prime; j++) {
         (*bucketArray)[j] = -1;
     }
 
     // Find in which bucket each element is and make the arrays for the bucket
     // and the chain
-    for (uint32_t i = startingPos + bucketSize - 1; i >= startingPos; i--) {
+    for (uint64_t i = startingPos + bucketSize - 1; i >= startingPos; i--) {
 
         // Get the hashed value
-        int32_t hashValue = h2(rel[i].value, prime);
+        uint64_t hashValue = h2(rel[i].value, prime);
 
         // If the corresponding bucket is currently empty
-        if ((*bucketArray)[hashValue] == (uint32_t) -1) {
+        if ((*bucketArray)[hashValue] == (uint64_t) -1) {
             (*bucketArray)[hashValue] = i - startingPos;
         }
         else {
             // If the bucket is not empty, we have a collision,
             // so we need to add a value in the chain array
-            int32_t chainPos = (*bucketArray)[hashValue];
-            while ((*chainArray)[chainPos] != (uint32_t) -1) {
+            uint64_t chainPos = (*bucketArray)[hashValue];
+            while ((*chainArray)[chainPos] != (uint64_t) -1) {
                 chainPos = (*chainArray)[chainPos];
             }
             (*chainArray)[chainPos] = i - startingPos;
@@ -100,7 +100,7 @@ void bucketify2(Tuple * rel,
 
     // Debugging prints
     // std::cout << "Ordered bucket:" << std::endl;
-    // for (uint32_t i = startingPos; i < bucketSize + startingPos; i++) {
+    // for (uint64_t i = startingPos; i < bucketSize + startingPos; i++) {
     //     std::cout << "\t"<<  i - startingPos  << ": "
     //     << rel->column[i].value << ": "
     //     << h2(rel->column[i].value, prime) << ": "
@@ -109,13 +109,13 @@ void bucketify2(Tuple * rel,
     // std::cout << std::endl;
 
     // std::cout << "bucketArray:" << std::endl;
-    // for (uint32_t i = 0; i < prime; i++) {
+    // for (uint64_t i = 0; i < prime; i++) {
     //     std::cout << "\t" << i << ": " << (*bucketArray)[i] << std::endl;
     // }
     // std::cout << std::endl;
 
     // std::cout << "chainArray:" << std::endl;
-    // for (uint32_t i = 0; i < bucketSize; i++) {
+    // for (uint64_t i = 0; i < bucketSize; i++) {
     //     std::cout << "\t" << i << ": " << (*chainArray)[i] << std::endl;
     // }
 }
