+++ join/join.hpp
@@ -4,12 +4,12 @@
 
 Result * join(Relation * A, Relation * B);
 
-void compare(Tuple * orderedBig, Tuple * orderedSmall, uint32_t bucketSizeSmall,\
-    uint32_t startIndexSmall, uint32_t bucketSizeBig, uint32_t startIndexBig, \
-    uint32_t * bucketArray, uint32_t * chainArray, Result * result, bool flag);
+void compare(Tuple * orderedBig, Tuple * orderedSmall, uint64_t bucketSizeSmall,\
+    uint64_t startIndexSmall, uint64_t bucketSizeBig, uint64_t startIndexBig, \
+    uint64_t * bucketArray, uint64_t * chainArray, Result * result, bool flag);
 
 void checkEquals(Tuple * valueA, int hash_value, Tuple * orderedSmall, \
-    uint32_t startIndexBig, uint32_t * bucketArray, uint32_t * chainArray, \
+    uint64_t startIndexBig, uint64_t * bucketArray, uint64_t * chainArray, \
     Result * result, bool flag, int rowIdBig);
 
 int naiveJoin(Relation * A, Relation * B);
