+++ join/structs.cpp
@@ -2,9 +2,9 @@
 
 // Create and return a Tuplearray with random values
 // Max value of random number is 9999
-Tuple * randomColumn(uint32_t size){
+Tuple * randomColumn(uint64_t size){
     Tuple * array = new Tuple[size];
-    for(uint32_t i = 0; i<size; i++){
+    for(uint64_t i = 0; i<size; i++){
         array[i].rowid = i+1;
         array[i].value = rand()%2000;
     }
@@ -12,7 +12,7 @@ Tuple * randomColumn(uint32_t size){
 }
 
 // Create and return a Relation with serial values
-Relation * randomRelation(uint32_t size){
+Relation * randomRelation(uint64_t size){
     Relation * rel = new Relation;
     rel->column = randomColumn(size);
     rel->size = size;
@@ -21,9 +21,9 @@ Relation * randomRelation(uint32_t size){
 
 
 // Create and return a Tuplearray with serial values
-Tuple * serialColumn(uint32_t size){
+Tuple * serialColumn(uint64_t size){
     Tuple * array = new Tuple[size];
-    for(uint32_t i = 0; i<size; i++){
+    for(uint64_t i = 0; i<size; i++){
         array[i].rowid = i+1;
         array[i].value = i+1;
     }
@@ -32,7 +32,7 @@ Tuple * serialColumn(uint32_t size){
 
 
 // Create and return a Relation with serial values
-Relation * serialRelation(uint32_t size){
+Relation * serialRelation(uint64_t size){
     Relation * rel = new Relation;
     rel->column = serialColumn(size);
     rel->size = size;
@@ -41,10 +41,10 @@ Relation * serialRelation(uint32_t size){
 
 
 // Create and return a Tuplearray with odd values
-Tuple * oddColumn(uint32_t size){
+Tuple * oddColumn(uint64_t size){
     Tuple * array = new Tuple[size];
-    int32_t value = 1;
-    for(uint32_t i = 0; i<size; i++){
+    uint64_t value = 1;
+    for(uint64_t i = 0; i<size; i++){
         array[i].rowid = i+1;
         array[i].value = value;
         value += 2;
@@ -54,7 +54,7 @@ Tuple * oddColumn(uint32_t size){
 
 
 // Create and return a Relation with odd values
-Relation * oddRelation(uint32_t size){
+Relation * oddRelation(uint64_t size){
     Relation * rel = new Relation;
     rel->column = oddColumn(size);
     rel->size = size;
@@ -63,10 +63,10 @@ Relation * oddRelation(uint32_t size){
 
 
 // Create and return a Tuplearray with even values
-Tuple * evenColumn(uint32_t size){
+Tuple * evenColumn(uint64_t size){
     Tuple * array = new Tuple[size];
-    int32_t value = 0;
-    for(uint32_t i = 0; i<size; i++){
+    uint64_t value = 0;
+    for(uint64_t i = 0; i<size; i++){
         array[i].rowid = i+1;
         array[i].value = value;
         value += 2;
@@ -75,7 +75,7 @@ Tuple * evenColumn(uint32_t size){
 }
 
 // Create and return a Relation with even values
-Relation * evenRelation(uint32_t size){
+Relation * evenRelation(uint64_t size){
     Relation * rel = new Relation;
     rel->column = evenColumn(size);
     rel->size = size;
@@ -89,7 +89,7 @@ void deleteRelation(Relation * rel){
 }
 
 void printRelation(Relation * rel){
-    for(uint32_t i=0; i<rel->size; i++){
+    for(uint64_t i=0; i<rel->size; i++){
         std::cout << rel->column[i].rowid << ". "
              << rel->column[i].value << std::endl;
     }
